requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-debug-toolbar==3.8.1
django-environ==0.10.0
django-filter==23.1
django-guardian==2.4.0
django-taggit==3.1.0
sqlparse==0.4.3
